rt: hide internal apis
