+++ ref/vk/vk_ray_internal.h
@@ -3,6 +3,7 @@
 #include "vk_core.h"
 #include "vk_buffer.h"
 #include "vk_const.h"
+#include "vk_rtx.h"
 
 #define MAX_INSTANCES 2048
 #define MAX_KUSOCHKI 32768
@@ -89,3 +90,44 @@ void XVK_RayModel_ClearForNextFrame( void );
 void XVK_RayModel_Validate(void);
 
 void RT_RayModel_Clear(void);
+
+// Just creates an empty BLAS structure, doesn't alloc anything
+// Memory pointed to by name must remain alive until RT_BlasDestroy
+struct rt_blas_s* RT_BlasCreate(const char *name, rt_blas_usage_e usage);
+
+// Create an empty BLAS with specified limits
+struct rt_blas_s* RT_BlasCreatePreallocated(const char *name, rt_blas_usage_e usage, int max_geometries, const int *max_prims, int max_vertex, uint32_t extra_buffer_offset);
+
+void RT_BlasDestroy(struct rt_blas_s* blas);
+
+// 1. Schedules BLAS build (allocates geoms+ranges from a temp pool, etc).
+// 2. Allocates kusochki (if not) and fills them with geom and initial material data
+qboolean RT_BlasBuild(struct rt_blas_s *blas, const struct vk_render_geometry_s *geoms, int geoms_count);
+
+VkDeviceAddress RT_BlasGetDeviceAddress(struct rt_blas_s *blas);
+
+typedef struct rt_kusochki_s {
+	uint32_t offset;
+	int count;
+	int internal_index__;
+} rt_kusochki_t;
+
+// TODO lifetime arg here is KORYAVY
+rt_kusochki_t RT_KusochkiAlloc(int count, r_geometry_lifetime_t lifetime);
+void RT_KusochkiFree(const rt_kusochki_t*);
+
+struct vk_render_geometry_s;
+qboolean RT_KusochkiUpload(const rt_kusochki_t *kusochki, const struct vk_render_geometry_s *geoms, int geoms_count, int override_texture_id);
+
+// Update animated materials
+void RT_KusochkiUploadSubset(rt_kusochki_t *kusochki, const struct vk_render_geometry_s *geoms, const int *geoms_indices, int geoms_indices_count);
+
+typedef struct {
+	const struct rt_blas_s* blas;
+	uint32_t kusochki_offset;
+	int render_type; // TODO material_mode
+	const matrix3x4 *transform, *prev_transform;
+	const vec4_t *color;
+} rt_blas_frame_args_t;
+
+void RT_BlasAddToFrame( rt_blas_frame_args_t args );
